+++ option.go
@@ -54,6 +54,13 @@ func WithLogHandler(handler slog.Handler) Option {
 	}
 }
 
+// WithOnStatus provides the callback for monitoring status of configuration watching.
+func WithOnStatus(onStatus func(loader Loader, changed bool, err error)) Option {
+	return func(options *options) {
+		options.onStatus = onStatus
+	}
+}
+
 type (
 	// Option configures a Config with specific options.
 	Option  func(*options)