+++ provider/appconfig/appconfig.go
@@ -14,7 +14,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"log/slog"
 	"reflect"
 	"sync/atomic"
 	"time"
@@ -30,13 +29,13 @@ import (
 type AppConfig struct {
 	unmarshal    func([]byte, any) error
 	pollInterval time.Duration
-	logger       *slog.Logger
 
-	client *clientProxy
+	onStatus func(bool, error)
+	client   *clientProxy
 }
 
 // New creates an AppConfig with the given application, environment, profile and Option(s).
-func New(application, environment, profile string, opts ...Option) AppConfig {
+func New(application, environment, profile string, opts ...Option) *AppConfig {
 	option := &options{
 		client: &clientProxy{
 			application: application,
@@ -50,16 +49,16 @@ func New(application, environment, profile string, opts ...Option) AppConfig {
 	option.client.pollInterval = option.pollInterval / 2 //nolint:gomnd
 	option.client.timeout = option.pollInterval / 2      //nolint:gomnd
 
-	return AppConfig(*option)
+	return (*AppConfig)(option)
 }
 
-func (a AppConfig) Load() (map[string]any, error) {
+func (a *AppConfig) Load() (map[string]any, error) {
 	values, _, err := a.load(context.Background())
 
 	return values, err
 }
 
-func (a AppConfig) Watch(ctx context.Context, onChange func(map[string]any)) error {
+func (a *AppConfig) Watch(ctx context.Context, onChange func(map[string]any)) error {
 	pollInterval := time.Minute
 	if a.pollInterval > 0 {
 		pollInterval = a.pollInterval
@@ -71,24 +70,9 @@ func (a AppConfig) Watch(ctx context.Context, onChange func(map[string]any)) err
 		select {
 		case <-ticker.C:
 			values, changed, err := a.load(ctx)
-			if err != nil {
-				logger := a.logger
-				if logger == nil {
-					logger = slog.Default()
-				}
-
-				logger.LogAttrs(
-					ctx, slog.LevelWarn,
-					"Error when reloading from AWS AppConfig",
-					slog.String("application", a.client.application),
-					slog.String("environment", a.client.environment),
-					slog.String("profile", a.client.profile),
-					slog.Any("error", err),
-				)
-
-				continue
+			if a.onStatus != nil {
+				a.onStatus(changed, err)
 			}
-
 			if changed {
 				onChange(values)
 			}
@@ -98,7 +82,7 @@ func (a AppConfig) Watch(ctx context.Context, onChange func(map[string]any)) err
 	}
 }
 
-func (a AppConfig) load(ctx context.Context) (map[string]any, bool, error) {
+func (a *AppConfig) load(ctx context.Context) (map[string]any, bool, error) {
 	resp, changed, err := a.client.load(ctx)
 	if !changed || err != nil {
 		return nil, false, err
@@ -116,7 +100,11 @@ func (a AppConfig) load(ctx context.Context) (map[string]any, bool, error) {
 	return values, true, nil
 }
 
-func (a AppConfig) String() string {
+func (a *AppConfig) Status(onStatus func(bool, error)) {
+	a.onStatus = onStatus
+}
+
+func (a *AppConfig) String() string {
 	return "appConfig:" + a.client.application + "-" + a.client.environment + "-" + a.client.profile
 }