+++ watch_test.go
@@ -111,6 +111,29 @@ func TestConfig_Watch_twice(t *testing.T) {
 	assert.Equal(t, expected, buf.String())
 }
 
+func TestConfig_Watch_status(t *testing.T) {
+	t.Parallel()
+
+	buf := &buffer{}
+	config := konf.New(konf.WithLogHandler(logHandler(buf)))
+	assert.NoError(t, config.Load(&statusWatcher{}))
+
+	stopped := make(chan struct{})
+	ctx, cancel := context.WithCancel(context.Background())
+	defer func() {
+		cancel()
+		<-stopped
+	}()
+	go func() {
+		defer close(stopped)
+		assert.NoError(t, config.Watch(ctx))
+	}()
+	time.Sleep(100 * time.Millisecond) // Wait for watch to start
+
+	expected := "level=WARN msg=\"Error when watching configuration changes.\" loader=status error=\"watch error\"\n"
+	assert.Equal(t, expected, buf.String())
+}
+
 func TestConfig_Watch_panic(t *testing.T) {
 	t.Parallel()
 
@@ -199,6 +222,28 @@ func (errorWatcher) Watch(context.Context, func(map[string]any)) error {
 
 func (errorWatcher) String() string {
 	return "error"
+}
+
+type statusWatcher struct {
+	onStatus func(bool, error)
+}
+
+func (*statusWatcher) Load() (map[string]any, error) {
+	return nil, nil //nolint:nilnil
+}
+
+func (s *statusWatcher) Watch(context.Context, func(map[string]any)) error {
+	s.onStatus(false, errors.New("watch error"))
+
+	return nil
+}
+
+func (s *statusWatcher) Status(onStatus func(bool, error)) {
+	s.onStatus = onStatus
+}
+
+func (*statusWatcher) String() string {
+	return "status"
 }
 
 func logHandler(buf *buffer) *slog.TextHandler {